usecase, dataManagerのロジックを修正
