+++ app/src/main/java/com/google/codelab/hotpepperapiapp/usecase/MapsUseCaseImpl.kt
@@ -3,6 +3,9 @@ package com.google.codelab.hotpepperapiapp.usecase
 import com.google.codelab.hotpepperapiapp.data.SearchDataManagerImpl
 import com.google.codelab.hotpepperapiapp.model.response.StoresResponse
 import io.reactivex.Single
+import io.reactivex.android.schedulers.AndroidSchedulers
+import io.reactivex.schedulers.Schedulers
+import java.io.IOException
 
 class MapsUseCaseImpl : MapsUseCase {
     private val dataManager: SearchDataManagerImpl = SearchDataManagerImpl()
@@ -12,5 +15,11 @@ class MapsUseCaseImpl : MapsUseCase {
         lng: Double
     ): Single<StoresResponse> {
         return dataManager.fetchStores(lat, lng)
+            .subscribeOn(Schedulers.io())
+            .observeOn(AndroidSchedulers.mainThread())
+            .map {
+                return@map it.body()
+                    ?: throw IOException("failed to fetch")
+            }
     }
 }
