+++ app/src/main/java/com/google/codelab/hotpepperapiapp/usecase/StoreListUseCaseImpl.kt
@@ -3,6 +3,9 @@ package com.google.codelab.hotpepperapiapp.usecase
 import com.google.codelab.hotpepperapiapp.data.SearchDataManagerImpl
 import com.google.codelab.hotpepperapiapp.model.response.StoresResponse
 import io.reactivex.Single
+import io.reactivex.android.schedulers.AndroidSchedulers
+import io.reactivex.schedulers.Schedulers
+import java.io.IOException
 
 class StoreListUseCaseImpl : StoreListUseCase {
     private val dataManager: SearchDataManagerImpl = SearchDataManagerImpl()
@@ -13,5 +16,11 @@ class StoreListUseCaseImpl : StoreListUseCase {
         start: Int
     ): Single<StoresResponse> {
         return dataManager.fetchStores(lat, lng, start)
+            .subscribeOn(Schedulers.io())
+            .observeOn(AndroidSchedulers.mainThread())
+            .map {
+                return@map it.body()
+                    ?: throw IOException("failed to fetch")
+            }
     }
 }
